@@ -2,7 +2,7 @@ name: local_auth
 description: Flutter plugin for Android and iOS device authentication sensors
   such as Fingerprint Reader and Touch ID.
 homepage: https://github.com/flutter/plugins/tree/master/packages/local_auth
-version: 0.6.3+5
+version: 1.0.0-nullsafety
 
 flutter:
   plugin:
@@ -16,10 +16,10 @@ flutter:
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.0.5
-  intl: ">=0.15.1 <0.17.0"
-  platform: ">=2.0.0 <4.0.0"
-  flutter_plugin_android_lifecycle: ^1.0.2
+  meta: ^1.3.0-nullsafety.3
+  intl: ^0.17.0-nullsafety.2
+  platform: ^3.0.0-nullsafety.4
+  flutter_plugin_android_lifecycle: ^2.0.0-nullsafety
 
 dev_dependencies:
   integration_test:
@@ -28,8 +28,8 @@ dev_dependencies:
     sdk: flutter
   flutter_test:
     sdk: flutter
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0-nullsafety.1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.12.13+hotfix.5"
