@@ -40,26 +40,26 @@ abstract class PathProviderPlatform extends PlatformInterface {
 
   /// Path to the temporary directory on the device that is not backed up and is
   /// suitable for storing caches of downloaded files.
-  Future<String> getTemporaryPath() {
+  Future<String?> getTemporaryPath() {
     throw UnimplementedError('getTemporaryPath() has not been implemented.');
   }
 
   /// Path to a directory where the application may place application support
   /// files.
-  Future<String> getApplicationSupportPath() {
+  Future<String?> getApplicationSupportPath() {
     throw UnimplementedError(
         'getApplicationSupportPath() has not been implemented.');
   }
 
   /// Path to the directory where application can store files that are persistent,
   /// backed up, and not visible to the user, such as sqlite.db.
-  Future<String> getLibraryPath() {
+  Future<String?> getLibraryPath() {
     throw UnimplementedError('getLibraryPath() has not been implemented.');
   }
 
   /// Path to a directory where the application may place data that is
   /// user-generated, or that cannot otherwise be recreated by your application.
-  Future<String> getApplicationDocumentsPath() {
+  Future<String?> getApplicationDocumentsPath() {
     throw UnimplementedError(
         'getApplicationDocumentsPath() has not been implemented.');
   }
@@ -67,7 +67,7 @@ abstract class PathProviderPlatform extends PlatformInterface {
   /// Path to a directory where the application may access top level storage.
   /// The current operating system should be determined before issuing this
   /// function call, as this functionality is only available on Android.
-  Future<String> getExternalStoragePath() {
+  Future<String?> getExternalStoragePath() {
     throw UnimplementedError(
         'getExternalStoragePath() has not been implemented.');
   }
@@ -76,7 +76,7 @@ abstract class PathProviderPlatform extends PlatformInterface {
   /// stored. These paths typically reside on external storage like separate
   /// partitions or SD cards. Phones may have multiple storage directories
   /// available.
-  Future<List<String>> getExternalCachePaths() {
+  Future<List<String>?> getExternalCachePaths() {
     throw UnimplementedError(
         'getExternalCachePaths() has not been implemented.');
   }
@@ -84,10 +84,10 @@ abstract class PathProviderPlatform extends PlatformInterface {
   /// Paths to directories where application specific data can be stored.
   /// These paths typically reside on external storage like separate partitions
   /// or SD cards. Phones may have multiple storage directories available.
-  Future<List<String>> getExternalStoragePaths({
+  Future<List<String>?> getExternalStoragePaths({
     /// Optional parameter. See [StorageDirectory] for more informations on
     /// how this type translates to Android storage directories.
-    StorageDirectory type,
+    StorageDirectory? type,
   }) {
     throw UnimplementedError(
         'getExternalStoragePaths() has not been implemented.');
@@ -95,7 +95,7 @@ abstract class PathProviderPlatform extends PlatformInterface {
 
   /// Path to the directory where downloaded files can be stored.
   /// This is typically only relevant on desktop operating systems.
-  Future<String> getDownloadsPath() {
+  Future<String?> getDownloadsPath() {
     throw UnimplementedError('getDownloadsPath() has not been implemented.');
   }
 }
